LPS-83359 Move common code to Base class
